+++ src/tools/ToolsWotLK/mmaps_generator/TerrainBuilder.cpp
@@ -187,7 +187,7 @@ namespace MMAP
             {
                 uint8 v9[V9_SIZE_SQ];
                 uint8 v8[V8_SIZE_SQ];
-                int count = 0;
+                size_t count = 0;
                 count += static_cast<int>(fread(v9, sizeof(uint8), V9_SIZE_SQ, mapFile));
                 count += static_cast<int>(fread(v8, sizeof(uint8), V8_SIZE_SQ, mapFile));
                 if (count != expected)
@@ -205,7 +205,7 @@ namespace MMAP
             {
                 uint16 v9[V9_SIZE_SQ];
                 uint16 v8[V8_SIZE_SQ];
-                int count = 0;
+                size_t count = 0;
                 count += static_cast<int>(fread(v9, sizeof(uint16), V9_SIZE_SQ, mapFile));
                 count += static_cast<int>(fread(v8, sizeof(uint16), V8_SIZE_SQ, mapFile));
                 if (count != expected)
@@ -221,7 +221,7 @@ namespace MMAP
             }
             else
             {
-                int count = 0;
+                size_t count = 0;
                 count += static_cast<int>(fread(V9, sizeof(float), V9_SIZE_SQ, mapFile));
                 count += static_cast<int>(fread(V8, sizeof(float), V8_SIZE_SQ, mapFile));
                 if (count != expected)
